requirements.txt
# =============================================================================
# ESG報告書提取器核心依賴包 v1.0
# =============================================================================

# 核心AI和機器學習框架
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=1.0.0
transformers>=4.40.0
sentence-transformers>=2.3.0
torch>=2.1.0

# 向量資料庫和搜索
faiss-cpu>=1.8.0

# 數據處理和科學計算
numpy>=1.24.0
pandas>=2.2.0
scikit-learn>=1.3.0

# PDF處理
pypdf>=4.0.0

# Excel文件處理
openpyxl>=3.1.0

# Google API相關
google-generativeai>=0.3.0
google-api-core>=2.11.0

# 系統和配置
python-dotenv>=1.0.0
tqdm>=4.66.0
pathlib

# 字符串相似度計算（用於公司名稱標準化）
difflib

# =============================================================================
# 安裝說明
# =============================================================================
# 
# 1. 建議使用虛擬環境：
#    python -m venv venv
#    source venv/bin/activate  # Linux/Mac
#    venv\Scripts\activate     # Windows
#
# 2. 安裝依賴：
#    pip install -r requirements.txt
#
# 3. 如果安裝torch時遇到問題，可以訪問：
#    https://pytorch.org/get-started/locally/
#    選擇適合您系統的版本
#
# 4. 如果需要GPU加速（可選）：
#    將 faiss-cpu 替換為 faiss-gpu
#    並安裝對應的CUDA版本的torch
#
# =============================================================================